requirements.txt
requests==2.23.0
pytelegrambotapi==3.6.7
pymongo==3.10.1
dependency_injector==3.15.6
jinja2==2.11.1
ciso8601
cachetools==4.0.0
flask==1.1.2
python_dotenv
dnspython
gunicorn